+++ core/include/EventHelper.h
@@ -126,9 +126,9 @@ private:
     Event::Handle<std::vector<GenJet>> h_genjets;
 
     Event::Handle<std::vector<bool>> h_triggerResults;
-    Event::Handle<std::vector<int>> h_triggerPrescales;
-    Event::Handle<std::vector<int>> h_triggerPrescalesL1min;
-    Event::Handle<std::vector<int>> h_triggerPrescalesL1max;
+    Event::Handle<std::vector<double>> h_triggerPrescales;
+    Event::Handle<std::vector<double>> h_triggerPrescalesL1min;
+    Event::Handle<std::vector<double>> h_triggerPrescalesL1max;
     Event::Handle<std::vector<std::string>> h_triggerNames;
     Event::Handle<std::vector<L1EGamma>> h_L1EG_seeds;
     Event::Handle<std::vector<L1Jet>> h_L1J_seeds;
